% File src/library/base/man/load.Rd
\name{vecapply}
\alias{vecapply}
\alias{va_veclen}
\title{Vectorization of apply family functions}
\description{
  Vectorization of apply family functions
}
\usage{
va_veclen(v)
}
\arguments{
  \item{v}{A vector input, could be atomic vector, matrix, array, or a listb}
}
\examples{
TBD
}
\keyword{apply}